Move inputs into forms folder
