+++ .storybook/decorators/UserContextDecorator.svelte
@@ -2,7 +2,7 @@
   import { setContext } from "svelte";
   import { writable, type Writable } from "svelte/store";
   import type { User } from "@/api/types";
-  import { me } from "@/api/fixtures/orgAndUser.fixtures";
+  import { me } from "@/test/fixtures/accounts.ts";
   const userStore: Writable<User> = writable(null);
   $: userStore.set(me as User);
   setContext("me", userStore);
